I Added the website assets ,all the images and SVG files.
Created an SVG.app file and i stored & made all the components that return SVG data ,i did this to make the app
less populated with long texts of code,i then exported the components to the App.js file and imported them there.

I learned that the svg tags on the SVG.js file needed a the viewBox & preserveAspectRatio attribute which i found on chartio.com.